Moving get content
Doing multiple commits incase i need to revert and cherry pick
